Refactor check if auto token is not present
